+++ ForestManagement/functions/acls/Unregister-FMAcl.ps1
@@ -1,0 +1,41 @@
+﻿function Unregister-FMAcl {
+	<#
+	.SYNOPSIS
+		Removes a acl that had previously been registered.
+	
+	.DESCRIPTION
+		Removes a acl that had previously been registered.
+	
+	.PARAMETER Path
+		The path (distinguishedName) of the acl to remove.
+
+	.PARAMETER Category
+		The object category the acl settings apply to
+	
+	.EXAMPLE
+		PS C:\> Get-FMAcl | Unregister-FMAcl
+
+		Clears all registered acls.
+	#>
+	
+	[CmdletBinding()]
+	param (
+		[Parameter(ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
+		[string[]]
+		$Path,
+
+		[Parameter(ValueFromPipelineByPropertyName = $true)]
+		[string[]]
+		$Category
+	)
+	
+	process {
+		foreach ($pathItem in $Path) {
+			if ($pathItem -eq '<default>') { $script:aclDefaultOwner = $null }
+			else { $script:acls.Remove($pathItem) }
+		}
+		foreach ($categoryItem in $Category) {
+			$script:aclByCategory.Remove($categoryItem)
+		}
+	}
+}